separate bridge fns & teleportation fn to separate service files
