+++ src/actions/networkAction.ts
@@ -18,6 +18,8 @@ import transactionService from 'services/transaction.service'
 import { createAction } from './createAction'
 import { BigNumberish } from 'ethers'
 import balanceService from 'services/balance.service'
+import bridgingService from 'services/bridging.service'
+import teleportationService from 'services/teleportation.service'
 
 export const fetchBalances = () =>
   createAction('BALANCE/GET', () => balanceService.getBalances())
@@ -35,12 +37,12 @@ export const fetchFastExits = () =>
   createAction('FASTEXITS/GETALL', () => transactionService.getFastExits())
 
 export const exitBOBA = (token: string, value: BigNumberish) =>
-  createAction('EXIT/CREATE', () => networkService.exitBOBA(token, value))
+  createAction('EXIT/CREATE', () => bridgingService.exitBOBA(token, value))
 
 //SWAP RELATED
 export const depositL1LP = (currency: string, value: BigNumberish) =>
   createAction('DEPOSIT/CREATE', () =>
-    networkService.depositL1LP(currency, value)
+    bridgingService.depositL1LP(currency, value)
   )
 
 export const isTeleportationOfAssetSupported = (
@@ -49,7 +51,11 @@ export const isTeleportationOfAssetSupported = (
   destChainId: string
 ) =>
   createAction('DEPOSIT/TELEPORTATION/TOKEN_SUPPORTED', () =>
-    networkService.isTeleportationOfAssetSupported(layer, asset, destChainId)
+    teleportationService.isTeleportationOfAssetSupported(
+      layer,
+      asset,
+      destChainId
+    )
   )
 
 export const depositWithLightBridge = (
@@ -59,22 +65,27 @@ export const depositWithLightBridge = (
   destChainId: BigNumberish
 ) =>
   createAction('DEPOSIT/CREATE', () =>
-    networkService.depositWithTeleporter(layer, currency, value, destChainId)
+    teleportationService.depositWithTeleporter(
+      layer,
+      currency,
+      value,
+      destChainId
+    )
   )
 
 //SWAP RELATED - Depositing into the L2LP triggers the swap-exit
 export const depositL2LP = (token: string, value: BigNumberish) =>
-  createAction('EXIT/CREATE', () => networkService.depositL2LP(token, value))
+  createAction('EXIT/CREATE', () => bridgingService.depositL2LP(token, value))
 
 //CLASSIC DEPOSIT ETH
 export const depositETHL2 = (payload) =>
   createAction('DEPOSIT/CREATE', () => {
-    return networkService.depositETHL2(payload)
+    return bridgingService.depositETHToL2(payload)
   })
 
 //DEPOSIT ERC20
 export const depositErc20 = (payload) =>
-  createAction('DEPOSIT/CREATE', () => networkService.depositErc20(payload))
+  createAction('DEPOSIT/CREATE', () => bridgingService.depositErc20(payload))
 
 export const approveERC20 = (
   value: BigNumberish,
@@ -83,7 +94,7 @@ export const approveERC20 = (
   contractABI?
 ) =>
   createAction('APPROVE/CREATE', () =>
-    networkService.approveERC20(
+    bridgingService.approveERC20(
       value,
       currency,
       approveContractAddress,
@@ -97,7 +108,7 @@ export const transfer = (
   currency: string
 ) =>
   createAction('TRANSFER/CREATE', () =>
-    networkService.transfer(recipient, value, currency)
+    bridgingService.transfer(recipient, value, currency)
   )
 
 export const fetchLookUpPrice = (params) =>
