+++ src/hooks/useBridge/index.ts
@@ -30,6 +30,7 @@ import {
   resetToken,
 } from 'actions/bridgeAction'
 import { INetwork, NetworkList } from '../../util/network/network.util'
+import teleportationService from 'services/teleportation.service'
 
 export const useBridge = () => {
   const dispatch = useDispatch<any>()
@@ -116,7 +117,8 @@ export const useBridge = () => {
       // step -1  approve token
       // step -2  deposit to Teleportation.
 
-      const { teleportationAddr } = networkService.getTeleportationAddress()
+      const { teleportationAddr } =
+        teleportationService.getTeleportationAddress()
       if (!teleportationAddr) {
         console.warn('Teleportation Address not available.')
         return
