+++ src/services/networkService.ts
@@ -747,732 +747,6 @@ class NetworkService {
     })
   }
 
-  // TODO: bridging.service.ts
-  //Move ETH from L1 to L2 using the standard deposit system
-  /******
-   * Deposit ETH from L1 to L2.
-   * Deposit ETH from L1 to another L2 account.
-   * */
-
-  async depositETHL2({ recipient = null, value_Wei_String }) {
-    try {
-      setFetchDepositTxBlock(false)
-
-      let depositTX
-      if (this.network === Network.ETHEREUM) {
-        if (!recipient) {
-          depositTX = await this.L1StandardBridgeContract!.connect(
-            this.provider!.getSigner()
-          ).depositETH(
-            this.L2GasLimit,
-            utils.formatBytes32String(new Date().getTime().toString()),
-            {
-              value: value_Wei_String,
-            }
-          )
-        } else {
-          depositTX = await this.L1StandardBridgeContract!.connect(
-            this.provider!.getSigner()
-          ).depositETHTo(
-            recipient,
-            this.L2GasLimit,
-            utils.formatBytes32String(new Date().getTime().toString()),
-            {
-              value: value_Wei_String,
-            }
-          )
-        }
-      } else {
-        if (!recipient) {
-          depositTX = await this.L1StandardBridgeContract!.connect(
-            this.provider!.getSigner()
-          ).depositNativeToken(
-            this.L2GasLimit,
-            utils.formatBytes32String(new Date().getTime().toString()),
-            {
-              value: value_Wei_String,
-            }
-          )
-        } else {
-          depositTX = await this.L1StandardBridgeContract!.connect(
-            this.provider!.getSigner()
-          ).depositNativeTokenTo(
-            recipient,
-            this.L2GasLimit,
-            utils.formatBytes32String(new Date().getTime().toString()),
-            {
-              value: value_Wei_String,
-            }
-          )
-        }
-      }
-
-      setFetchDepositTxBlock(true)
-
-      //at this point the tx has been submitted, and we are waiting...
-      await depositTX.wait()
-
-      const opts = {
-        fromBlock: -4000,
-      }
-
-      const receipt = await this.watcher!.waitForMessageReceipt(depositTX, opts)
-      const txReceipt = receipt.transactionReceipt
-      console.log('completed Deposit! L2 tx hash:', receipt.transactionReceipt)
-      return txReceipt
-    } catch (error) {
-      console.log('NS: depositETHL2 error:', error)
-      return error
-    }
-  }
-
-  // TODO: bridging.service.ts
-  // REVIEW: More over looks like deprecated so we can remove this.
-  //Transfer funds from one account to another, on the L2
-  async transfer(
-    address: string,
-    value_Wei_String: BigNumberish,
-    currency: string
-  ) {
-    let tx: TransactionResponse
-
-    try {
-      if (currency === this.addresses.L2_ETH_Address) {
-        //we are sending ETH
-
-        const wei = BigNumber.from(value_Wei_String)
-
-        tx = await this.provider!.getSigner().sendTransaction({
-          to: address,
-          value: ethers.utils.hexlify(wei),
-        })
-      } else {
-        //any ERC20 json will do....
-        tx = await this.L2_TEST_Contract!.connect(this.provider!.getSigner())
-          .attach(currency)
-          .transfer(address, value_Wei_String)
-        await tx.wait()
-      }
-
-      return tx
-    } catch (error) {
-      console.log('NS: transfer error:', error)
-      return error
-    }
-  }
-
-  // TODO: bridging.service.ts
-  // REVIEW: More over looks like deprecated so we can remove this.
-  //Estimate funds transfer from one account to another, on the L2
-  async transferEstimate(
-    recipient: string,
-    value_Wei_String: string,
-    currency: string
-  ) {
-    const gasPrice_BN = await this.L2Provider!.getGasPrice()
-
-    let cost_BN = BigNumber.from('0')
-    let gas_BN = BigNumber.from('0')
-
-    try {
-      if (currency === this.addresses.L2_ETH_Address) {
-        gas_BN = await this.provider!.getSigner().estimateGas({
-          from: this.account,
-          to: recipient,
-          value: value_Wei_String,
-        })
-
-        cost_BN = gas_BN.mul(gasPrice_BN)
-        console.log('ETH: Transfer cost in ETH:', utils.formatEther(cost_BN))
-      } else {
-        const ERC20Contract = new ethers.Contract(
-          currency,
-          L2StandardERC20ABI, // any old abi will do...
-          this.provider!.getSigner()
-        )
-
-        const tx = await ERC20Contract.populateTransaction.transfer(
-          recipient,
-          value_Wei_String
-        )
-
-        gas_BN = await this.L2Provider!.estimateGas(tx)
-
-        cost_BN = gas_BN.mul(gasPrice_BN)
-        console.log('ERC20: Transfer cost in ETH:', utils.formatEther(cost_BN))
-      }
-
-      const safety_margin = BigNumber.from('1000000000000')
-      console.log('ERC20: Safety margin:', utils.formatEther(safety_margin))
-
-      return cost_BN.add(safety_margin)
-    } catch (error) {
-      console.log('NS: transferEstimate error:', error)
-      return error
-    }
-  }
-
-  async checkAllowance(currencyAddress: string, targetContract: string) {
-    console.log('currencyAddress', currencyAddress)
-    console.log('targetContract', targetContract)
-    try {
-      const ERC20Contract = new ethers.Contract(
-        currencyAddress,
-        L1ERC20ABI, //could use any abi - just something with .allowance
-        this.provider!.getSigner()
-      )
-      return await ERC20Contract.allowance(this.account, targetContract)
-    } catch (error) {
-      console.log('NS: checkAllowance error:', error)
-      return error
-    }
-  }
-
-  async approveERC20(
-    value_Wei_String: BigNumberish,
-    currency: string,
-    approveContractAddress: string = this.addresses.L1StandardBridgeAddress,
-    contractABI = L1ERC20ABI
-  ) {
-    try {
-      const ERC20Contract = new ethers.Contract(
-        currency,
-        contractABI,
-        this.provider!.getSigner()
-      )
-
-      /***********************/
-
-      let allowance_BN = await ERC20Contract.allowance(
-        this.account,
-        approveContractAddress
-      )
-      console.log('Initial Allowance is:', allowance_BN)
-
-      /*
-      OMG IS A SPECIAL CASE - allowance needs to be set to zero, and then
-      set to actual amount, unless current approval amount is equal to, or
-      bigger than, the current approval value
-      */
-      if (
-        this.networkGateway === Network.ETHEREUM &&
-        allowance_BN.lt(BigNumber.from(value_Wei_String)) &&
-        currency.toLowerCase() === allTokens.OMG.L1.toLowerCase()
-      ) {
-        console.log(
-          "Current OMG Token allowance too small - might need to reset to 0, unless it's already zero"
-        )
-        if (allowance_BN.gt(BigNumber.from('0'))) {
-          const approveOMG = await ERC20Contract.approve(
-            approveContractAddress,
-            ethers.utils.parseEther('0')
-          )
-          await approveOMG.wait()
-          console.log('OMG Token allowance has been set to 0')
-        }
-      }
-
-      //recheck the allowance
-      allowance_BN = await ERC20Contract.allowance(
-        this.account,
-        approveContractAddress
-      )
-      console.log('Second Allowance is:', allowance_BN)
-
-      const allowed = allowance_BN.gte(BigNumber.from(value_Wei_String))
-
-      console.log('Allowed?:', allowed)
-
-      if (!allowed) {
-        console.log('Not good enough - need to set to:', value_Wei_String)
-        //and now, the normal allowance transaction
-        const approveStatus = await ERC20Contract.approve(
-          approveContractAddress,
-          value_Wei_String
-        )
-        await approveStatus.wait()
-        console.log('ERC20 L1 SWAP ops approved:', approveStatus)
-      }
-
-      return true
-    } catch (error) {
-      console.log('NS: approveERC20 error:', error)
-      return error
-    }
-  }
-
-  // TODO: bridging.service.ts
-  //Used to move ERC20 Tokens from L1 to L2 using the classic deposit
-  async depositErc20({
-    recipient = null,
-    value_Wei_String,
-    currency,
-    currencyL2,
-  }) {
-    const L1_TEST_Contract = this.L1_TEST_Contract!.attach(currency)
-
-    let allowance_BN = await L1_TEST_Contract.allowance(
-      this.account,
-      this.addresses.L1StandardBridgeAddress
-    )
-    setFetchDepositTxBlock(false)
-    try {
-      /*
-      OMG IS A SPECIAL CASE - allowance needs to be set to zero, and then
-      set to actual amount, unless current approval amount is equal to, or
-      bigger than, the current approval value
-      */
-      if (
-        this.networkGateway === Network.ETHEREUM &&
-        allowance_BN.lt(BigNumber.from(value_Wei_String)) &&
-        currency.toLowerCase() === allTokens.OMG.L1.toLowerCase()
-      ) {
-        console.log(
-          "Current OMG Token allowance too small - might need to reset to 0, unless it's already zero"
-        )
-        if (allowance_BN.gt(BigNumber.from('0'))) {
-          const approveOMG = await L1_TEST_Contract.approve(
-            this.addresses.L1StandardBridgeAddress,
-            ethers.utils.parseEther('0')
-          )
-          await approveOMG.wait()
-          console.log('OMG Token allowance has been set to 0')
-        }
-      }
-
-      //recheck the allowance
-      allowance_BN = await L1_TEST_Contract.allowance(
-        this.account,
-        this.addresses.L1StandardBridgeAddress
-      )
-
-      const allowed = allowance_BN.gte(BigNumber.from(value_Wei_String))
-
-      if (!allowed) {
-        //and now, the normal allowance transaction
-        const approveStatus = await L1_TEST_Contract.connect(
-          this.provider!.getSigner()
-        ).approve(this.addresses.L1StandardBridgeAddress, value_Wei_String)
-        await approveStatus.wait()
-        console.log('ERC20 L1 ops approved:', approveStatus)
-      }
-      let depositTX
-      if (!recipient) {
-        // incase no recipient
-        depositTX = await this.L1StandardBridgeContract!.connect(
-          this.provider!.getSigner()
-        ).depositERC20(
-          currency,
-          currencyL2,
-          value_Wei_String,
-          this.L2GasLimit,
-          utils.formatBytes32String(new Date().getTime().toString())
-        )
-      } else {
-        // deposit ERC20 to L2 account address.
-        depositTX = await this.L1StandardBridgeContract!.connect(
-          this.provider!.getSigner()
-        ).depositERC20To(
-          currency,
-          currencyL2,
-          recipient,
-          value_Wei_String,
-          this.L2GasLimit,
-          utils.formatBytes32String(new Date().getTime().toString())
-        )
-      }
-      setFetchDepositTxBlock(true)
-      //at this point the tx has been submitted, and we are waiting...
-      await depositTX.wait()
-
-      const opts = {
-        fromBlock: -4000,
-      }
-      const receipt = await this.watcher!.waitForMessageReceipt(depositTX, opts)
-      const txReceipt = receipt.transactionReceipt
-      await balanceService.getBalances()
-      return txReceipt
-    } catch (error) {
-      console.log('NS: depositErc20 error:', error)
-      return error
-    }
-  }
-
-  // TODO: bridging.service.ts
-  //Standard 7 day exit from BOBA
-  async exitBOBA(currencyAddress, value_Wei_String) {
-    await updateSignatureStatus_exitTRAD(false)
-
-    try {
-      const L2BillingContract = new ethers.Contract(
-        this.addresses.Proxy__BobaBillingContract,
-        L2BillingContractABI,
-        this.L2Provider
-      )
-      let BobaApprovalAmount = await L2BillingContract.exitFee()
-
-      //now coming in as a value_Wei_String
-      const value = BigNumber.from(value_Wei_String)
-
-      const allowance = await this.checkAllowance(
-        currencyAddress,
-        this.addresses.DiscretionaryExitFee
-      )
-
-      const BobaAllowance = await this.checkAllowance(
-        this.addresses.TK_L2BOBA,
-        this.addresses.DiscretionaryExitFee
-      )
-
-      if (this.networkGateway === Network.ETHEREUM) {
-        // Should approve BOBA
-        if (
-          utils.getAddress(currencyAddress) ===
-          utils.getAddress(this.addresses.TK_L2BOBA)
-        ) {
-          BobaApprovalAmount = BobaApprovalAmount.add(value)
-        }
-
-        if (BobaAllowance.lt(BobaApprovalAmount)) {
-          const res = await this.approveERC20(
-            BobaApprovalAmount,
-            this.addresses.TK_L2BOBA,
-            this.addresses.DiscretionaryExitFee
-          )
-          if (!res) {
-            return false
-          }
-        }
-      }
-
-      let otherField
-      if (this.networkGateway === Network.ETHEREUM) {
-        otherField =
-          currencyAddress === this.addresses.L2_ETH_Address ? { value } : {}
-      } else {
-        otherField =
-          currencyAddress === this.addresses.L2_ETH_Address
-            ? { value: value.add(BobaApprovalAmount) }
-            : { value: BobaApprovalAmount }
-      }
-
-      // Should approve other tokens
-      if (
-        currencyAddress !== this.addresses.L2_ETH_Address &&
-        utils.getAddress(currencyAddress) !==
-          utils.getAddress(this.addresses.TK_L2BOBA) &&
-        allowance.lt(value)
-      ) {
-        const res = await this.approveERC20(
-          value,
-          currencyAddress,
-          this.addresses.DiscretionaryExitFee
-        )
-        if (!res) {
-          return false
-        }
-      }
-
-      const DiscretionaryExitFeeContract = new ethers.Contract(
-        this.addresses.DiscretionaryExitFee,
-        DiscretionaryExitFeeABI,
-        this.provider!.getSigner()
-      )
-
-      const tx = await DiscretionaryExitFeeContract.payAndWithdraw(
-        currencyAddress,
-        value_Wei_String,
-        this.L1GasLimit,
-        utils.formatBytes32String(new Date().getTime().toString()),
-        otherField
-      )
-
-      //everything submitted... waiting
-      await tx.wait()
-
-      //can close window now
-      await updateSignatureStatus_exitTRAD(true)
-
-      return tx
-    } catch (error) {
-      console.log('NS: exitBOBA error:', error)
-      return error
-    }
-  }
-
-  /***********************************************************/
-  /***** SWAP ON to BOBA by depositing funds to the L1LP *****/
-  /***********************************************************/
-
-  // TODO: move to bridging service as it's corresponds to fast deposits.
-  async depositL1LP(currency, value_Wei_String) {
-    try {
-      await updateSignatureStatus_depositLP(false)
-      setFetchDepositTxBlock(false)
-
-      const depositTX = await this.L1LPContract!.connect(
-        this.provider!.getSigner()
-      ).clientDepositL1(
-        value_Wei_String,
-        currency,
-        currency === this.addresses.L1_ETH_Address
-          ? { value: value_Wei_String }
-          : {}
-      )
-
-      setFetchDepositTxBlock(true)
-
-      //at this point the tx has been submitted, and we are waiting...
-      await depositTX.wait()
-      await updateSignatureStatus_depositLP(true)
-
-      const opts = {
-        fromBlock: -4000,
-      }
-      const receipt = await this.watcher!.waitForMessageReceipt(depositTX, opts)
-      const txReceipt = receipt.transactionReceipt
-      console.log(' completed swap-on ! L2 tx hash:', txReceipt)
-      return txReceipt
-    } catch (error) {
-      console.log('NS: depositL1LP error:', error)
-      return error
-    }
-  }
-
-  // TODO: teleportation.service.ts
-  getTeleportationAddress(chainId?: number) {
-    if (!chainId) {
-      chainId = this.chainId
-    }
-    const networkConfig = CHAIN_ID_LIST[chainId!]
-    let teleportationAddr
-    if (!networkConfig) {
-      console.error(
-        `Unknown chainId to retrieve teleportation contract from: ${chainId}`
-      )
-      return { teleportationAddr: null, networkConfig: null }
-    }
-    const addresses = appService.fetchAddresses({
-      networkType: networkConfig.networkType,
-      network: networkConfig.chain,
-    })
-    if (networkConfig.layer === LAYER.L1) {
-      teleportationAddr = addresses.Proxy__L1Teleportation
-    } else if (networkConfig.layer === LAYER.L2) {
-      teleportationAddr = addresses.Proxy__L2Teleportation
-    }
-    return { teleportationAddr, networkConfig }
-  }
-
-  // TODO: teleportation.service.ts
-  getTeleportationContract(chainId) {
-    const { teleportationAddr, networkConfig } =
-      this.getTeleportationAddress(chainId)
-    if (!teleportationAddr || !this.Teleportation) {
-      return
-    }
-
-    const rpc = getRpcUrl({
-      networkType: networkConfig.networkType,
-      network: networkConfig.chain,
-      layer: networkConfig.layer,
-    })
-    const provider = new ethers.providers.StaticJsonRpcProvider(rpc)
-
-    return this.Teleportation!.attach(teleportationAddr).connect(provider)
-  }
-
-  // TODO: teleportation.service.ts
-  async isTeleportationOfAssetSupported(layer, token, destChainId) {
-    const teleportationAddr =
-      layer === Layer.L1
-        ? this.addresses.Proxy__L1Teleportation
-        : this.addresses.Proxy__L2Teleportation
-    if (!teleportationAddr) {
-      return { supported: false }
-    }
-    const contract = this.Teleportation!.attach(teleportationAddr).connect(
-      this.provider!.getSigner()
-    )
-    return contract.supportedTokens(token, destChainId)
-  }
-
-  // TODO: teleportation.service.ts
-  async depositWithTeleporter(layer, currency, value_Wei_String, destChainId) {
-    try {
-      updateSignatureStatus_depositLP(false)
-      setFetchDepositTxBlock(false)
-
-      const teleportationAddr =
-        layer === Layer.L1
-          ? this.addresses.Proxy__L1Teleportation
-          : this.addresses.Proxy__L2Teleportation
-      const msgVal =
-        currency === this.addresses.L1_ETH_Address ||
-        currency === this.addresses.NETWORK_NATIVE
-          ? { value: value_Wei_String }
-          : {}
-      const teleportationContract = this.Teleportation!.attach(
-        teleportationAddr
-      ).connect(this.provider!.getSigner())
-      const tokenAddress =
-        currency === this.addresses.NETWORK_NATIVE
-          ? ethers.constants.AddressZero
-          : currency
-
-      const assetSupport = await teleportationContract.supportedTokens(
-        tokenAddress,
-        destChainId
-      )
-      if (!assetSupport?.supported) {
-        console.error(
-          'Teleportation: Asset not supported for chainId',
-          assetSupport,
-          tokenAddress,
-          destChainId
-        )
-        return new Error(
-          `Teleportation: Asset ${tokenAddress} not supported for chainId ${destChainId}`
-        )
-      }
-
-      const depositTX = await teleportationContract.teleportAsset(
-        tokenAddress,
-        value_Wei_String,
-        destChainId,
-        msgVal
-      )
-
-      setFetchDepositTxBlock(true)
-
-      //at this point the tx has been submitted, and we are waiting...
-      await depositTX.wait()
-      updateSignatureStatus_depositLP(true)
-
-      const opts = {
-        fromBlock: -4000,
-      }
-      const receipt = await this.watcher!.waitForMessageReceipt(depositTX, opts)
-      const txReceipt = receipt.transactionReceipt
-      console.log(' completed swap-on ! tx hash:', txReceipt)
-      return txReceipt
-    } catch (error) {
-      console.log('Teleportation error:', error)
-      return error
-    }
-  }
-
-  /**************************************************************/
-  /***** SWAP OFF from BOBA by depositing funds to the L2LP *****/
-  /**************************************************************/
-  // TODO: bridging.service.ts
-  async depositL2LP(currencyAddress: string, value_Wei_String: BigNumberish) {
-    await updateSignatureStatus_exitLP(false)
-
-    console.log('depositL2LP currencyAddress', currencyAddress)
-
-    const L2BillingContract = new ethers.Contract(
-      this.addresses.Proxy__BobaBillingContract,
-      L2BillingContractABI,
-      this.L2Provider
-    )
-    let BobaApprovalAmount = await L2BillingContract.exitFee()
-
-    const BobaAllowance = await this.checkAllowance(
-      this.addresses.TK_L2BOBA,
-      this.addresses.L2LPAddress
-    )
-
-    try {
-      if (this.networkGateway === Network.ETHEREUM) {
-        // Approve BOBA first only when the Boba is not native token.
-        if (
-          utils.getAddress(currencyAddress) ===
-          utils.getAddress(this.addresses.TK_L2BOBA)
-        ) {
-          BobaApprovalAmount = BobaApprovalAmount.add(
-            BigNumber.from(value_Wei_String)
-          )
-        }
-        if (BobaAllowance.lt(BobaApprovalAmount)) {
-          const approveStatus = await this.approveERC20(
-            BobaApprovalAmount,
-            this.addresses.TK_L2BOBA,
-            this.addresses.L2LPAddress
-          )
-          if (!approveStatus) {
-            return false
-          }
-        }
-      }
-
-      // Approve other tokens
-      if (
-        currencyAddress !== this.addresses.L2_ETH_Address &&
-        utils.getAddress(currencyAddress) !==
-          utils.getAddress(this.addresses.TK_L2BOBA)
-      ) {
-        const L2ERC20Contract = new ethers.Contract(
-          currencyAddress,
-          L2StandardERC20ABI,
-          this.provider!.getSigner()
-        )
-
-        const allowance_BN = await L2ERC20Contract.allowance(
-          this.account,
-          this.addresses.L2LPAddress
-        )
-
-        const depositAmount_BN = BigNumber.from(value_Wei_String)
-
-        if (depositAmount_BN.gt(allowance_BN)) {
-          const approveStatus = await L2ERC20Contract.approve(
-            this.addresses.L2LPAddress,
-            value_Wei_String
-          )
-          await approveStatus.wait()
-          if (!approveStatus) {
-            return false
-          }
-        }
-      }
-
-      const time_start = new Date().getTime()
-      console.log('TX start time:', time_start)
-
-      let otherField
-      if (this.networkGateway === Network.ETHEREUM) {
-        otherField =
-          currencyAddress === this.addresses.L2_ETH_Address
-            ? { value: value_Wei_String }
-            : {}
-      } else {
-        otherField =
-          currencyAddress === this.addresses.L2_ETH_Address
-            ? { value: BobaApprovalAmount.add(value_Wei_String) }
-            : { value: BobaApprovalAmount }
-      }
-
-      const depositTX = await this.L2LPContract!.connect(
-        this.provider!.getSigner()
-      ).clientDepositL2(value_Wei_String, currencyAddress, otherField)
-
-      //at this point the tx has been submitted, and we are waiting...
-      await depositTX.wait()
-
-      const block = await this.L2Provider!.getTransaction(depositTX.hash)
-      console.log(' block:', block)
-
-      //closes the modal
-      await updateSignatureStatus_exitLP(true)
-
-      return depositTX
-    } catch (error) {
-      console.log('NS: depositL2LP error:', error)
-      return error
-    }
-  }
-
   // TODO: price.service.ts or api.service.ts
   async fetchLookUpPrice(params) {
     try {
